open-source-pdf-qa/requirements.txt
# Python=3.10
python-dotenv
langchain
huggingface_hub
sentence-transformers
PyPDF2
chainlit
cohere
qdrant-client
rank_bm25
llama-cpp-python
